Refactor NoteFragment parsing code to NoteFragmentFeatureUtil.
