+++ src/main/java/seedu/address/logic/parser/ParserUtil.java
@@ -255,94 +255,4 @@ public class ParserUtil {
         return tagSet;
     }
 
-    /**
-     * Parses a {@code Note note} into a {@code List<NoteFragment>}.
-     * @param note The Note to parse.
-     * @return Returns a List of {@code NoteFragments}.
-     */
-    public static List<NoteFragment> parseNoteFragmentsFromNote(Note note) {
-        try {
-            Title title = note.getTitle();
-            List<String> noteFragmentMatches = parseNoteFragmentMatches(note.getContent().toString());
-            return parseNoteFragmentsFromString(title, noteFragmentMatches);
-        } catch (ParseException e) {
-            return Collections.emptyList();
-        }
-    }
-
-    /**
-     * Parses all {@code NoteFragment} patterns from a String.
-     * @param noteContent The content of the Note, as a String.
-     * @return Returns a List of matching Strings.
-     */
-    public static List<String> parseNoteFragmentMatches(String noteContent) {
-        noteContent = " " + noteContent;
-        ArgumentMultimap argMultimap = ArgumentTokenizer.tokenize(noteContent, PREFIX_NOTE_FRAGMENT_START,
-                PREFIX_NOTE_FRAGMENT_END);
-
-        List<String> noteFragmentList = argMultimap.getAllValues(PREFIX_NOTE_FRAGMENT_START);
-        return noteFragmentList;
-    }
-
-    /**
-     * Parses a {@code List<String>} into a {@code List<NoteFragment>}.
-     * @param title The {@code Title} of the {@code Note} this {@code List<String>} was derived from.
-     * @param stringList A List of matching pattern Strings from a {@code Note}.
-     * @return Returns a {@code List<NoteFragment>} containing the {@code NoteFragments} of a {@code Note}.
-     * @throws ParseException A custom Exception.
-     */
-    public static List<NoteFragment> parseNoteFragmentsFromString(Title title, List<String> stringList)
-            throws ParseException {
-        List<NoteFragment> noteFragmentList = new ArrayList<>();
-        for (String string : stringList) {
-            string = " " + string;
-            ArgumentMultimap argMultimapCheck = ArgumentTokenizer.tokenize(string, PREFIX_NOTE_FRAGMENT_CONTENT,
-                    PREFIX_NOTE_FRAGMENT_TAG);
-
-            if (!arePrefixesPresent(argMultimapCheck, PREFIX_NOTE_FRAGMENT_CONTENT, PREFIX_NOTE_FRAGMENT_TAG)) {
-                throw new ParseException(MESSAGE_INCORRECT_NOTE_FRAGMENT_FORMAT);
-            }
-
-            noteFragmentList.add(new NoteFragment(title, parseContentFromNoteFragment(string),
-                    parseTagsFromNoteFragment(string)));
-        }
-        return noteFragmentList;
-
-    }
-
-    /**
-     * Parses a {@code Content} from a {@code NoteFragment}.
-     * @param noteFragment The input {@code NoteFragment} object to parse.
-     * @return Returns a {@code Content} object.
-     * @throws ParseException A custom exception.
-     */
-    public static Content parseContentFromNoteFragment(String noteFragment) throws ParseException {
-        ArgumentMultimap argMultimap = ArgumentTokenizer.tokenize(noteFragment, PREFIX_NOTE_FRAGMENT_CONTENT,
-                PREFIX_NOTE_FRAGMENT_TAG);
-        try {
-            return parseContent(argMultimap.getValue(PREFIX_NOTE_FRAGMENT_CONTENT).get());
-        } catch (ParseException e) {
-            throw new ParseException(Content.MESSAGE_CONSTRAINTS);
-        }
-    }
-
-    /**
-     * Parses {@code Set<Tag> tags} from a {@code NoteFragment}.
-     * @param noteFragment The input {@code NoteFragment} object to parse.
-     * @return Returns a {@code Set<Tag>} set of {@code tags}.
-     * @throws ParseException A custom exception.
-     */
-    public static Set<Tag> parseTagsFromNoteFragment(String noteFragment) throws ParseException {
-        ArgumentMultimap argMultimap = ArgumentTokenizer.tokenize(noteFragment, PREFIX_NOTE_FRAGMENT_CONTENT,
-                PREFIX_NOTE_FRAGMENT_TAG);
-        try {
-            return parseTags(argMultimap.getAllValues(PREFIX_NOTE_FRAGMENT_TAG));
-        } catch (ParseException e) {
-            throw new ParseException(Tag.MESSAGE_CONSTRAINTS);
-        }
-    }
-
-    private static boolean arePrefixesPresent(ArgumentMultimap argumentMultimap, Prefix... prefixes) {
-        return Stream.of(prefixes).allMatch(prefix -> argumentMultimap.getValue(prefix).isPresent());
-    }
 }
